requirements.txt
Flask==3.1.3
flask-cors==6.0.2
google-generativeai==0.8.6
Pillow==12.1.1
python-dotenv==1.2.1
requests==2.32.0
gunicorn==23.0.0
